implemented user type filter and moved files accordingly
